Move ImageRenderer
