+++ src/components/HTMLEngineProvider/BaseHTMLEngineProvider.js
@@ -1,5 +1,4 @@
 import React, {useMemo} from 'react';
-import {TouchableOpacity} from 'react-native';
 import {
     TRenderEngineProvider,
     RenderHTMLConfigProvider,
@@ -9,11 +8,9 @@ import PropTypes from 'prop-types';
 import AnchorRenderer from './HTMLRenderers/AnchorRenderer';
 import CodeRenderer from './HTMLRenderers/CodeRenderer';
 import EditedRenderer from './HTMLRenderers/EditedRenderer';
-import Config from '../../CONFIG';
+import ImageRenderer from './HTMLRenderers/ImageRenderer';
 import styles from '../../styles/styles';
 import fontFamily from '../../styles/fontFamily';
-import AttachmentModal from '../AttachmentModal';
-import ThumbnailImage from '../ThumbnailImage';
 
 const propTypes = {
     /** Whether text elements should be selectable */
@@ -56,65 +53,6 @@ function computeEmbeddedMaxWidth(tagName, contentWidth) {
     return contentWidth;
 }
 
-function ImgRenderer(props) {
-    const htmlAttribs = props.tnode.attributes;
-
-    // There are two kinds of images that need to be displayed:
-    //
-    //     - Chat Attachment images
-    //
-    //           Images uploaded by the user via the app or email.
-    //           These have a full-sized image `htmlAttribs['data-expensify-source']`
-    //           and a thumbnail `htmlAttribs.src`. Both of these URLs need to have
-    //           an authToken added to them in order to control who
-    //           can see the images.
-    //
-    //     - Non-Attachment Images
-    //
-    //           These could be hosted from anywhere (Expensify or another source)
-    //           and are not protected by any kind of access control e.g. certain
-    //           Concierge responder attachments are uploaded to S3 without any access
-    //           control and thus require no authToken to verify access.
-    //
-    const isAttachment = Boolean(htmlAttribs['data-expensify-source']);
-    const originalFileName = htmlAttribs['data-name'];
-    let previewSource = htmlAttribs.src;
-    let source = isAttachment
-        ? htmlAttribs['data-expensify-source']
-        : htmlAttribs.src;
-
-    // Update the image URL so the images can be accessed depending on the config environment
-    previewSource = previewSource.replace(
-        Config.EXPENSIFY.URL_EXPENSIFY_COM,
-        Config.EXPENSIFY.URL_API_ROOT,
-    );
-    source = source.replace(
-        Config.EXPENSIFY.URL_EXPENSIFY_COM,
-        Config.EXPENSIFY.URL_API_ROOT,
-    );
-
-    return (
-        <AttachmentModal
-            sourceURL={source}
-            isAuthTokenRequired={isAttachment}
-            originalFileName={originalFileName}
-        >
-            {({show}) => (
-                <TouchableOpacity
-                    style={styles.noOutline}
-                    onPress={() => show()}
-                >
-                    <ThumbnailImage
-                        previewSourceURL={previewSource}
-                        style={styles.webViewStyles.tagStyles.img}
-                        isAuthTokenRequired={isAttachment}
-                    />
-                </TouchableOpacity>
-            )}
-        </AttachmentModal>
-    );
-}
-
 // Declare nonstandard tags and their content model here
 const customHTMLElementModels = {
     edited: defaultHTMLElementModels.span.extend({
@@ -133,7 +71,7 @@ const customHTMLElementModels = {
 const renderers = {
     a: AnchorRenderer,
     code: CodeRenderer,
-    img: ImgRenderer,
+    img: ImageRenderer,
     edited: EditedRenderer,
 };
